Create README.md
This repository contains an implementation of DPO-SSAC, as described in the corresponding paper, integrated with Convlab-3.
DPO-SSAC is an off-policy deep reinforcement learning method designed for task-oriented dialogue policy learning. It aims to enhance the stability, exploration, robustness, and learning speed of dialogue policies in discrete task-oriented dialogue environments.
One of the key features of DPO-SSAC is the introduction of the sigmoidal discrete-SAC (SSAC) approach, which addresses underestimation bias that may lead to pessimistic exploration in discrete-SAC, particularly in the context of dialogue systems.